requirements.txt
Django==1.5.3
Fabric==1.7.0
Pillow==2.1.0
South==0.8.2
boto==2.11.0
dj-database-url==0.2.2
django-appconf==0.6
django-compressor==1.3
django-debug-toolbar==0.9.4
django-heroku-memcacheify==0.4
django-mptt==0.5.5
django-polymorphic==0.5.2
django-pylibmc-sasl==0.2.4
django-storages==1.1.8
django-suit==0.2.5
django-taggit==0.10
easy-thumbnails==1.4
facebook-sdk==0.4.0
gunicorn==18.0
oauthlib==0.6.0
paramiko==1.11.0
psycopg2==2.5.1
pycrypto==2.6
pylibmc==1.2.3
requests==2.0.0
requests-oauth==0.4.1
requests-oauthlib==0.4.0
six==1.4.1
twitter-text-python==1.0.1
wsgiref==0.1.2
